Travis: Initiating build for commit 'b2638d9e9ce07b7d0c594ad4450ae53c75c600e8' on branch 'cleanerDatatables': refs #4985 Improve Dashboard / Datatables UI
refs #4985 Improve Dashboard / Datatables UI

various UI fixes and cleaner datatables

some more ui tweaks

this should make sure to related report text is centered when not displayed within dashboard

darker arrow icon

if there are no footer icons, do nothing

some fixes for the ui improvments, especially annotations related

there should be still a gap when not displayed in dashboard

save the expanded footer state to make sure users who often use the footer do not have to open it each time and also makes sure the footer will be still expanded after reloading the widget etc

move related report closer to the actual report

make sure all annotation pins are in one line and positioned correct

fix issues when dashboard widget is maximised. Annotations were not positioned correct and drawer was displayed

when there is no data or no paging, do not display paging or search

display all websites dashboard similar to other reports

do not display column documentation in dashboard but in other reports

show title on drawer hover

hide related reports if there are no related reports, makes the footer looks better when widgetized

make sure the annotations are positioned correct if opened within a widgetized report transition

a fix for firefox and make sure annotation pins do not jump in dashboard area when loading message is displayed

only on firefox and only within the first two columns the arrow was not displayed fully although the css of the image and parent nodes is the same. this workaround makes sure the arrow of the drawer is always fully visible on firefox

move annotations 6px to the left to be displayed as before