Travis: Initiating build for commit 'e263aba19adb3c3e74feae65439968b465372e0a' on branch 'code-refactorings': Refactorings in ScheduledReports to remove useless indentation
I have refactored the code a bit to make it (hopefully) more readable. I have replaced this:

```php
if (self::manageEvent($reportType)) {
    // do something
}
```

By this:

```php
if (! self::manageEvent($reportType)) {
    return;
}

// do something
```

It's not much but it removes one level of indentation in almost the whole class, which is kind of nice.

FYI, this is one of the "object calisthenics" rules.