Travis: Initiating build for commit '1eb64a40a0c5220bfe40273de33c0a130980ad99' on branch 'master': refs #6341 always calculate the remaining visitor cookie lifetime.
Otherwise the behavior whether lifetime will be extended or whether
remaining will be calculated would depend on the time at which
setVisitorCookieTimeout is called which is very confusing. Also makes
sure in case the page is open for an hour and there is a tracking
request the lifetime will not be extended by an hour.