Travis: Initiating build for commit '320b0f19ef4b1cbe4fffaba5488a4a2d518313fd' on branch '6617': refs #6617 if process or output size is too large, declare it as finished
Declares a process as finished as soon as a PID file is > 500bytes
(contains only PID) and output file is > 100MB (which should usually
not be larger than 100KB or 1MB) to prevent files growing up to many
GBs. I added unit and integration tests for the file size detection. It is
now possible to mock the methods file_exists and filesize although very
simple so far. Later we can allow to define callbacks or to define different
return values for different files or we can use something like vfsStream