Travis: Initiating build for commit 'a2d438198f171e0844318e1bec45c0bc7573e1a2' on branch '6728': refs #6728 automatically cache settings in tracker cache if used in tracker mode.
As soon as a setting is used in tracker mode we will cache it now making it much
easier for developers to use settings in tracker mode and we will no longer
query the database on each tracking request if a setting is requested apart from
cache creation. Refactored the settings API under the hood to make this possible.
The classes were doing things they were not supposed to do. For instance settings
class also took care of storage which it should not and settings class took
care of many validation things that should be done in a setting class. I cleaned
up that code and it is also much better testable now.