Travis: Initiating build for commit '40d4e1e54cb3a197829249cde8bd580785dfaa5d' on branch 'master': refs #6022 added option to edit goal on goal page.
Added new attribute edit-url to enriched headline. Setting this will
make the headline clickable, on hover appears a title and pointer to
indicate that it is clickable and to explain what happens when clicking
on it. I think this is more intuitive than showing an edit icon on hover
which would be an alternative. As it is done in enriched headline we
could otherwise change the behaviour easily.

Also fixed feature name for goals no longer contains the actual goal
name.