Travis: Initiating build for commit '1a64ebcd70d6851ee37053da5e9b620118f8e067' on branch '4633': refs #4633 a couple of fixes for one click segmentation
Trying to hack segment value into core (for each datatable row)
but this is basically impossible. We will most likely need a solution
for each report separately in case it is possible at all. Also a
few more features and bugfixes for the overlay and detection when
we can use oneClickSegment and when not